Added draft title, authors and citation
Until now, we used the default JOSS paper, to check the GitHub Actions
build process.

With this commit, we bring the original author names and title of
paper in main body of paper. Also, we add a sentence for each section
as a placeholder and added a citation. Finally, we added the source of
GNU GPL v3+.

This was done with the help of Mohammad Akhlaghi.