Edits to the author list and summary
Until now, the author list was based on the initial stage of the
project, but as it has evolved, it was necessary to update it. Also,
the summary needed some editing.

With this commit, Elham has been brought to the second author status
because of her very complete work on the tutorial. Giulia has also
been brought up because of finding bugs.