Correct some parts of paper and add two new section
Until now, I did not mentioned Gnuastro conisit of how many section
and did not exist any section for the developer.

With this commit, I explain about more details of Gnuastro and explain
for the user how report their bugs and debugging it.

Thanks to Mohammad Akhlaghi for his helping.